Fixed eslint of components folder
