+++ .prettierrc
@@ -1,4 +1,5 @@
 {
 	"tabWidth": 4,
-	"trailingComma": "es5"
+	"trailingComma": "es5",
+	"semi": true
 }
